User soft delete

+++ app/Http/Controllers/UserController.php
@@ -1,7 +1,7 @@
 <?php
-    
+
 namespace App\Http\Controllers;
-    
+
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\User;
@@ -10,7 +10,7 @@ use DB;
 use Hash;
 use Illuminate\Support\Arr;
 use Validator;
-    
+
 class UserController extends Controller
 {
     /**
@@ -19,9 +19,12 @@ class UserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index(Request $request){
-        $data = User::orderBy('id','ASC')->paginate(5);
+        $itemsOnPage = 10;
+        
+        //$data = User::orderBy('id','ASC')->paginate($itemsOnPage);
+        $data = User::withTrashed()->orderBy('deleted_at','ASC')->orderBy('username','ASC')->orderBy('id','ASC')->paginate($itemsOnPage); // withTrashed() also shows deleted 
         return view('users.index',compact('data'))
-            ->with('i', ($request->input('page', 1) - 1) * 5);
+            ->with('i', ($request->input('page', 1) - 1) * $itemsOnPage);
     }
     
     /**
@@ -40,8 +43,7 @@ class UserController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
+    public function store(Request $request){
         $this->validate($request, [
             'username' => 'required|unique:users,username',
             'name' => 'required',
@@ -50,13 +52,17 @@ class UserController extends Controller
             'password' => 'required|same:confirm-password',
             'roles' => 'required'
         ]);
-    
+        
+        if ($this->checkUsername($request->username)){
+            return redirect()->back()->withErrors(['username' => 'User with same username already exists'])->withInput();
+        }
+        
         $input = $request->all();
         $input['password'] = Hash::make($input['password']);
     
         $user = User::create($input);
         $user->assignRole($request->input('roles'));
-    
+        
         return redirect()->route('users.index')
                         ->with('success','User created successfully');
     }
@@ -102,14 +108,18 @@ class UserController extends Controller
             'password' => 'same:confirm-password',
             'roles' => 'required'
         ]);
-    
+        
+        if ($this->checkUsername($request->username, $id)){
+            return redirect()->back()->withErrors(['username' => 'User with same username already exists'])->withInput();
+        }
+        
         $input = $request->all();
         if(!empty($input['password'])){ 
             $input['password'] = Hash::make($input['password']);
         }else{
-            $input = Arr::except($input,array('password'));    
+            $input = Arr::except($input,array('password'));
         }
-    
+        
         $user = User::find($id);
         $user->update($input);
         DB::table('model_has_roles')->where('model_id',$id)->delete();
@@ -126,8 +136,42 @@ class UserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id){
-        User::find($id)->delete();
+        $userToDelete = User::find($id);
+        $userToDelete->update(array('username' => 'deleted_'.$userToDelete->username));
+        $userToDelete->delete();
         return redirect()->route('users.index')
-                        ->with('success','User deleted successfully');
+                        ->with('success','User (soft) deleted successfully');
+    }
+    
+    /**
+     * Restore the specified user from trash.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function restore($id){
+        $userToRestore = User::onlyTrashed()->find($id);
+        $usernameClean = substr($userToRestore->username, strlen('deleted_'));
+        
+        if ($this->checkUsername($usernameClean)){
+            
+            return back()->withErrors(['duplicatedUser' => 'User with same username already exists']);
+        }
+        
+        $userToRestore->update(array('username' => $usernameClean));
+        $userToRestore->restore();
+        return redirect()->route('users.index')
+                        ->with('success','User restored successfully');
+    }
+    
+    // CHECK if user with same username exists (case INsensitive)
+    private function checkUsername($username, $id = NULL){
+        $lowTitle = array_map("strtolower", [$username]);
+        $qery = "SELECT COUNT(*) as stevilka FROM users WHERE LOWER(users.username) = '".$lowTitle[0]."'";
+        if($id != NULL){
+            $qery = $qery."AND users.id != ".$id."";
+        }
+        $stevilo = DB::select( DB::raw( $qery));
+        return $stevilo[0]->stevilka > 0;
     }
 }
